Fix smooth scrolling problem caused by int rounding
Smooth scrolling now uses qreal

+++ src/canvas.h
@@ -81,8 +81,8 @@ private:
 	GotoLine *goto_line;
 	QLabel *page_overlay;
 
-	int mx, my;
-	int mx_down, my_down;
+	qreal mx, my;
+	qreal mx_down, my_down;
 	bool triple_click_possible;
 
 	int hide_mouse_timeout;
